Started to work on some testing.

+++ CGameEngine/src/Engine/Util/Util.cpp
@@ -8,12 +8,12 @@ void ENGINE::UTIL::splitString(const std::string &s, char delim, std::vector<std
 	}
 }
 
-float ENGINE::UTIL::colorpartToOneRange(unsigned int cPart)
+float ENGINE::UTIL::colorpartToOneRange(uint8_t cPart)
 {
 	return ((float)cPart)/255.0f;
 }
 
-glm::vec3 ENGINE::UTIL::colorToRGB(unsigned int r, unsigned int g, unsigned int b)
+glm::vec3 ENGINE::UTIL::colorToRGB(uint8_t r, uint8_t g, uint8_t b)
 {
 	return glm::vec3(colorpartToOneRange(r), colorpartToOneRange(g), colorpartToOneRange(b));
 }
