+++ CGameEngine/src/Engine/Util/Util.h
@@ -8,8 +8,8 @@
 #include <iostream>
 namespace ENGINE {
 	namespace UTIL {
-		float colorpartToOneRange(unsigned int cPart);
-		glm::vec3 colorToRGB(unsigned int r, unsigned int g, unsigned int b);
+		float colorpartToOneRange(uint8_t cPart);
+		glm::vec3 colorToRGB(uint8_t r, uint8_t g, uint8_t b);
 
 		std::vector<std::string> splitString(const std::string &s, char delim);
 		void splitString(const std::string & s, char delim, std::vector<std::string>& elems);
